requirements.txt
beautifulsoup4
fastapi
python-dotenv
requests
uvicorn
gunicorn
pymongo
